Move salt.utils.human_to_bytes to salt.utils.stringutils
